Ignore dotenvy errors in macro
dotenvy parses .env files differently than e.g. docker, failing if it
encounters a quote in the middle of a value, but populates the
environment with items it had already parsed even if it errors. This
combination allows for a usage where the variables dotenvy has to parse
can be put at the top of the .env file, and any additional variables can
be put below.

We do that internally to avoid having two .env files with identical
values (one for sqlx-cli and sqlx-macro, and one for populating a docker
environment). To achieve this we use a custom dotenv parser emulating
dockers behaviour (just complete enough for our usage, and currently not
open sourced) to read our .env file. We then put the variables sqlx-cli
and sqlx-macro needs at the top of the file. This though currently
doesn't work, as dotenvy error.